requirements.txt
aiogram~=2.25.1
peewee
python-dotenv~=1.0.0
requests~=2.31.0
pathlib~=1.0.1
docx2txt~=0.8
pypdf2~=3.0.1